+++ awc/CHANGES.md
@@ -8,6 +8,7 @@
 ### Changed
 * Feature `cookies` is now optional and enabled by default. [#1981]
 * `ClientBuilder::connector` method would take `actix_http::client::Connector<T, U>` type. [#2008]
+* Basic auth password now takes blank passwords as an empty string instead of Option. [#2050]
 
 ### Removed
 * `ClientBuilder::default` function [#2008]
@@ -17,6 +18,8 @@
 [#1981]: https://github.com/actix/actix-web/pull/1981
 [#2008]: https://github.com/actix/actix-web/pull/2008
 [#2024]: https://github.com/actix/actix-web/pull/2024
+[#2050]: https://github.com/actix/actix-web/pull/2050
+
 
 ## 3.0.0-beta.2 - 2021-02-10
 ### Added